server/requirements-base.txt
# TODO: launchpad lib
numpy
scipy
DmxPy
# PyQt5
# pyqtgraph
aubio
pyyaml